+++ package.json
@@ -34,5 +34,8 @@
     "js-yaml": "^3.12.0",
     "json-refs": "^3.0.10",
     "serverless-domain-manager": "^2.6.5"
+  },
+  "dependencies": {
+    "fs-extra": "^7.0.1"
   }
 }
